+++ src/components/mobile/survey-questions.tsx
@@ -27,7 +27,7 @@ import { RadioGroup, RadioGroupItem } from "~/components/ui/radio-group";
 
 import { Card, CardContent, CardHeader, CardTitle } from "~/components/ui/card";
 import { type useForm } from "react-hook-form";
-import { Button } from "../ui/button";
+import { Button } from "~/components/ui/button";
 
 export function MobileSurveyQuestionnaire({
     userId,
